Backend/requirements.txt
python-dotenv>=0.19.0
fastapi
mangum
uvicorn
autogen-core
autogen-agentchat
autogen-ext[openai]
autogen-core
aiohttp>=3.8.0
requests>=2.28.0
Pillow>=9.0.0
#pyautogen>=0.2.0
boto3>=1.26.0
aws-lambda-powertools>=2.8.0
